Implemented an ImageBuilder
Impemented a few classes to build very specific image based on selected
stacks. This allows for example the building of a projection of from
a specified range of slices with or without background subtraction.

Also implemented two more command line options, -r and -c for specifying
a file containing the colors to be used, and one for the viewing ranges
to use for setting the bias and scale (contrast & brightness).

Didn't do much with the cavity mapper function. Hopefully around the
next commit.

+++ image/background.cpp
@@ -62,10 +62,10 @@ float Background::bg(int x, int y, int z){
 	cerr << "Background::bg point outsde area: " << x << "," << y << "," << z << endl;
 	return(0);
     }
-    // this is why I don't like unsigned ints.
-    int xb = ((int)x - (int)x_m/2) / (int)x_m;
-    int yb = ((int)y - (int)y_m/2) / (int)y_m;
-    int zb = ((int)z - (int)z_m/2) / (int)z_m;
+
+    int xb = (x - (int)x_m/2) / (int)x_m;
+    int yb = (y - (int)y_m/2) / (int)y_m;
+    int zb = (z - (int)z_m/2) / (int)z_m;
     
     // if x cannot be negative, then, x - x_m/2 >= -x_m/2
     // this means that the smallest number will be -0.5, which
